When loading skills from the cache treat metadata as constant
The hard-coded type, label and so forth of hard-coded skills
should not be overwritten from the cache.

This way we don't have to purge or manually update the cache when,
say, the label or noedit/noroll flags or whatever of a predefined
skill change.

Base and skill value are always loaded from the cache though.