Skills may be 0 or -ve and still rollable
Debuffs/penalties/etc may reduce a skill to 0 or below:
Such a skill is still usable, but only succeeds on a 1.

Skills which are naturally 0 (a base score of 0 and no
learned level) should still be unusable, but that will
have to be implemented elsewhere as the information has
been discarded by the time roll_d100 is invoked.