Abstraction over the storage interface to allow separate datasets
Use the document's URI to generate a prefix from the ?query part
of the URI, if any. This is then added to, and stripped from,
all storage access requsts as appropriate, allowing different
base URIs to result in different datasets, effectively allowing
multiple caracters to be stored at the same protocol+server+port
end point.