Melee and missile weapon skills re-templated
These skills now also record the weapon category, since
all weapons in a category may be used at 50% of the best
skill in said category.

We also store the stat requirements as these will be used to
debuff the skill when breached (TODO).

We also store the damage value as we want to automate
damage rolling, esp for specials & criticals.